requirements_wsl.txt
absl-py==1.4.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==23.1.2
cloudpickle==2.2.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.1.0
cryptography==2.8
cycler==0.11.0
Cython==0.29.36
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fonttools==4.41.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.56.0
h5py==2.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.8.0
importlib-resources==6.0.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lvis==0.5.3
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.2.1.16
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.2.53
nvidia-cudnn-cu11==8.9.2.26
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-dali-cuda110==1.20.0
nvidia-dali-nvtf-plugin==1.20.0+nv22.12
nvidia-horovod==0.26.1+nv22.12
nvidia-nccl-cu11==2.18.3
nvidia-pyindex==1.0.9
nvidia-tensorflow==1.15.5+nv22.12
oauthlib==3.1.0
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pexpect==4.6.0
Pillow==8.4.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycocotools==2.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2023.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.4
ssh-import-id==5.10
systemd-python==234
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorrt==8.6.1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
tf-slim==1.1.0
Twisted==18.9.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.1
zope.interface==4.7.1
